Precision loss when converting from double to float on iOS 4.3.
A hickup filter was added.

+++ machete/MacheteGLView.h
@@ -23,7 +23,7 @@ extern CAEAGLLayer *eaglLayer;
 
 @interface MacheteGLView : UIView {
   machete::graphics::DrawContext* renEngine;
-  float timestamp;
+  double timestamp;
     
   machete::IGame *game;
 }
